requirements.txt
numpy==1.15.2
torch==1.0.0.dev20180929
onnx==1.3.0
Pillow==5.3.0
torchvision==0.2.1
